# Reverted poll interval when waiting request status