<!-- Describe the changes in this PR -->

Increase the poll interval in https://github.com/skypilot-org/skypilot/pull/6889 introduce high latency for short requests, should consider other potential solution to reduce the CPU usage.

Master: 

```
multitime -n 5 sky status
            Mean        Std.Dev.    Min         Median      Max
real        3.889       0.163       3.656       3.831       4.117
user        1.084       0.136       0.881       1.122       1.277
sys         0.743       0.197       0.518       0.670       1.103
```

This PR:

```
multitime -n 5 sky status
1: sky status
            Mean        Std.Dev.    Min         Median      Max
real        1.610       0.261       1.353       1.459       2.025
user        0.841       0.085       0.775       0.785       1.000
sys         0.811       0.151       0.618       0.763       0.993
```

Before #6889

```
multitime -n 5 sky status
1: sky status
            Mean        Std.Dev.    Min         Median      Max
real        1.981       0.894       1.251       1.598       3.702
user        1.023       0.216       0.857       0.922       1.438
sys         0.737       0.093       0.658       0.674       0.880
```

<!-- Describe the tests ran -->
<!-- Unit tests (tests/test_*.py) are part of GitHub CI; below are tests that launch on the cloud. -->

Tested (run the relevant ones):

- [ ] Code formatting: install pre-commit (auto-check on commit) or `bash format.sh`
- [ ] Any manual or new tests for this PR (please specify below)
- [ ] All smoke tests: `/smoke-test` (CI) or `pytest tests/test_smoke.py` (local)
- [ ] Relevant individual tests: `/smoke-test -k test_name` (CI) or `pytest tests/test_smoke.py::test_name` (local)
- [ ] Backward compatibility: `/quicktest-core` (CI) or `pytest tests/smoke_tests/test_backward_compat.py` (local)

<!-- CI commands (/-prefixed) can only be triggered by repo members -->